menambah pop up delete

+++ application/views/dinasluar/dl_view.php
@@ -52,7 +52,7 @@
 							<td align="left">
 								<a title = "Cetak DL" class="btn btn-danger" href="<?php echo base_url() ?>DinasLuar/cetak/<?php echo $data['id_dl'];?>"><span class="glyphicon glyphicon-file"></span></a>
 								<a title = "Edit DL" class="btn btn-danger " href="<?php echo base_url() ?>DinasLuar/update/<?php echo $data['id_dl'];?>"><span class="glyphicon glyphicon-edit"></span></a>
-								<a title = "Delete DL" class="btn btn-danger btn-xs btn-flat" href="<?php echo base_url() ?>DinasLuar/delete/<?php echo $data['id_dl'];?>"><span class="glyphicon glyphicon-trash"></span></a>
+								<a title = "Delete DL" class="btn btn-danger btn-xs btn-flat" onclick="return confirmDialog()"  href="<?php echo base_url() ?>DinasLuar/delete/<?php echo $data['id_dl'];?>"><span class="glyphicon glyphicon-trash"></span></a>
 							</td>
 						</tr>
